SpendWise/pyproject.toml
[project]
name = "spendwise"
version = "0.1.0"
description = "SpendWise - AI-powered spending analysis with MCP sandbox"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.3",
    "python-dotenv>=1.0.0",
    "mcp-sandbox-openai-sdk",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "twilio>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.uv.sources]
mcp-sandbox-openai-sdk = { path = "../python-mcp-sandbox-openai-sdk", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spendwise"]

[tool.ruff]
line-length = 100
